Fixes commands exchange alghorithm
Changes the commands response package - it sends the 'from' position from
client history with list.
Fixes the loop that was bind to container size, that is modified in this loop.
Fixes const end in draw history, because hash was computed for all commands.
Moves registering operators to more suitable place.
Changes list registered type to pair.